+++ src/Plugins/RpcClient/Models/RpcApplicationLog.cs
@@ -9,6 +9,7 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
+using Neo.Extensions;
 using Neo.Json;
 using Neo.SmartContract;
 using Neo.VM;
